fix a bunch of tests

+++ src/py21cmfast/src/_functionprototypes_wrapper.h
@@ -9,7 +9,7 @@ int ComputePerturbField(float redshift,  UserParams *user_params,  CosmoParams *
 
 int ComputeHaloField(float redshift_desc, float redshift,  UserParams *user_params,  CosmoParams *cosmo_params,
                       AstroParams *astro_params,  FlagOptions *flag_options,
-                      InitialConditions *boxes, int random_seed,  HaloField * halos_desc,  HaloField *halos);
+                      InitialConditions *boxes, unsigned long long int random_seed,  HaloField * halos_desc,  HaloField *halos);
 
 int ComputePerturbHaloField(float redshift,  UserParams *user_params,  CosmoParams *cosmo_params,
                       AstroParams *astro_params,  FlagOptions *flag_options,
